requirements.txt
Flask
openai>=1.0.0
python-dotenv
requests
Pillow
PyMuPDF
